🏷️ fix(vote): fix incorrect parameter type and name

+++ src/main/java/com/ptjcoding/nbcampspringnewsfeed/domain/vote/repository/interfaces/VoteJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VoteJpaRepository extends JpaRepository<VoteEntity, Integer> {
+public interface VoteJpaRepository extends JpaRepository<VoteEntity, Long> {
 
 }
